refactor: split repository to packages
