+++ src/app.py
@@ -31,7 +31,6 @@ app.include_router(auth_controller.router, tags=["auth"], prefix="/auth")
 app.include_router(user_controller.router, tags=["user"], prefix="/user")
 app.include_router(tenant_controller.router, tags=["tenant"], prefix="/tenant")
 app.include_router(mode_controller.router, tags=["mode"], prefix="/mode")
-# app.include_router(config_controller.router, tags=["config"], prefix="/config")
 app.include_router(prompt_controller.router, tags=["prompt"], prefix="/prompt")
 app.include_router(chat_controller.router, tags=["chat"], prefix="/chat")
 app.include_router(message_controller.router, tags=["message"], prefix="/message")
