+++ src/model/file/file_model.py
@@ -14,7 +14,7 @@ from datastep.chains.datastep_file_description_chain import get_chain
 from datastep.components import datastep_faiss, datastep_multivector
 from exception.file_not_found_exception import FileNotFoundException
 from infra.redis_queue import get_redis_queue, QueueName, MAX_JOB_TIMEOUT
-from repository import file_repository
+from repository.file import file_repository
 from scheme.file.file_scheme import File, DataExtract, FileCreate
 from scheme.nomenclature.nomenclature_scheme import JobIdRead
 from util.files_paths import get_file_folder_path, get_file_storage_path
